Create QLab 5 - Level Adjust - Specific Output
This script will change volume on all CURRENTLY PLAYING tracks if they are routed to the outputs defined below. Will only work in edit mode.
See below for user variables, such as whether or not you want to set an absolute level or a relative level, and what the level will be,
You should put this script into a script cue and trigger it with a hotkey from a Stream Deck, or give it a unique number and trigger it from Companion.